requirements.txt
ccxt
requests
ttkbootstrap
Pillow
python-dateutil
tabulate
aiohttp
pyyaml
ruamel.yaml
numpy
async_timeout